Modify the jackson.js and move the token extraction needs to move into jackson.js
